all: Move object! macro to graph crate
