+++ graphql/src/store/resolver.rs
@@ -4,7 +4,7 @@ use std::result;
 use std::sync::Arc;
 
 use graph::components::store::*;
-use graph::data::graphql::ObjectOrInterface;
+use graph::data::graphql::{object, ObjectOrInterface};
 use graph::prelude::*;
 
 use crate::prelude::*;
